moved webpack configs to configs folder
