+++ configs/webpack.config.base.js
@@ -5,19 +5,23 @@
 import path from 'path';
 import webpack from 'webpack';
 import fs from 'fs';
-import { dependencies as externals } from './app/package';
-import { dependencies as possibleExternals } from './package';
+import { dependencies as externals } from '../app/package';
+import { dependencies as possibleExternals } from '../package';
 
 // Find all the dependencies without a `main` property and add them as webpack externals
 function filterDepWithoutEntryPoints(dep: string): boolean {
   // Return true if we want to add a dependency to externals
   try {
     // If the root of the dependency has an index.js, return true
-    if (fs.existsSync(path.join(__dirname, `node_modules/${dep}/index.js`))) {
+    if (
+      fs.existsSync(path.join(__dirname, '..', `node_modules/${dep}/index.js`))
+    ) {
       return false;
     }
     const pgkString = fs
-      .readFileSync(path.join(__dirname, `node_modules/${dep}/package.json`))
+      .readFileSync(
+        path.join(__dirname, '..', `node_modules/${dep}/package.json`)
+      )
       .toString();
     const pkg = JSON.parse(pgkString);
     const fields = ['main', 'module', 'jsnext:main', 'browser'];
@@ -50,7 +54,7 @@ export default {
   },
 
   output: {
-    path: path.join(__dirname, 'app'),
+    path: path.join(__dirname, '..', 'app'),
     // https://github.com/webpack/webpack/issues/1114
     libraryTarget: 'commonjs2'
   },
@@ -60,7 +64,7 @@ export default {
    */
   resolve: {
     extensions: ['.js', '.jsx', '.json'],
-    modules: [path.join(__dirname, 'app'), 'node_modules']
+    modules: [path.join(__dirname, '..', 'app'), 'node_modules']
   },
 
   plugins: [
