+++ configs/webpack.config.main.prod.js
@@ -2,12 +2,13 @@
  * Webpack config for production electron main process
  */
 
+import path from 'path';
 import webpack from 'webpack';
 import merge from 'webpack-merge';
 import UglifyJSPlugin from 'uglifyjs-webpack-plugin';
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
 import baseConfig from './webpack.config.base';
-import CheckNodeEnv from './internals/scripts/CheckNodeEnv';
+import CheckNodeEnv from '../internals/scripts/CheckNodeEnv';
 
 CheckNodeEnv('production');
 
@@ -21,7 +22,7 @@ export default merge.smart(baseConfig, {
   entry: './app/main.dev',
 
   output: {
-    path: __dirname,
+    path: path.join(__dirname, '..'),
     filename: './app/main.prod.js'
   },
