+++ configs/webpack.config.renderer.dev.dll.js
@@ -8,15 +8,15 @@ import webpack from 'webpack';
 import path from 'path';
 import merge from 'webpack-merge';
 import baseConfig from './webpack.config.base';
-import { dependencies } from './package';
-import CheckNodeEnv from './internals/scripts/CheckNodeEnv';
+import { dependencies } from '../package';
+import CheckNodeEnv from '../internals/scripts/CheckNodeEnv';
 
 CheckNodeEnv('development');
 
-const dist = path.resolve(process.cwd(), 'dll');
+const dist = path.join(__dirname, '..', 'dll');
 
 export default merge.smart(baseConfig, {
-  context: process.cwd(),
+  context: path.join(__dirname, '..'),
 
   devtool: 'eval',
 
@@ -66,9 +66,9 @@ export default merge.smart(baseConfig, {
     new webpack.LoaderOptionsPlugin({
       debug: true,
       options: {
-        context: path.resolve(process.cwd(), 'app'),
+        context: path.join(__dirname, '..', 'app'),
         output: {
-          path: path.resolve(process.cwd(), 'dll')
+          path: path.join(__dirname, '..', 'dll')
         }
       }
     })
